pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "wiki-acronym-tables"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "requests",
    "openpyxl",
    "pyyaml",
]

[project.optional-dependencies]
dev = ["pytest"]

[project.scripts]
wiki-acronym-tables = "wiki_acronyms.cli:main"
wiki-poetry = "wiki_acronyms.poetry_cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
